[FileProcessing] Compare resources implementation
With this change we will now be able to run a comparison
of image resources. The code is syncronized, as we are
dealing with multiple threads, for keeping track of what
have been done.

If dimensions do not match then the comparison will skip
the iterated pair.

If matches are found then, depending on the percentage,
they will either be labeled as "duplicates" (0% difference)
or as "matching" (<10% difference). Respective resources
will be moved to a directory with said names, once the
entire comparison is done.

Logging will be performed so a history of duplicates/matching
will be preserved, if we want to end the scan earlier.

Note: For larger collections of images it is recommended
to use more than +50 threads as each image is compared with
eachother, pixel by pixel. The work of moving files, that are
either a duplicate or matching, will be performed at the end
of the execution, hence the amount of threads will make a
difference on the execution time.