[FileProcessing] Crop resources implementation
With this change we will now be able to crop image
resources. The code is syncronized, as we are dealing
with multiple threads, to provide logging of what have
been done so it can be read afterwards.

As files are cropped they are moved right away to the
"done" directory, with this feature we aren't waiting
until the end of the execution.

Note: This implementation expect the borders to be of
equal sizes on the Y and the X axis. Do not use this
feature unless this is the case, as you will loose part
of the image as a result.