[FileProcessing] Convert resources implementation
With this change we will now be able to convert image
resources from one extension to another that are supported.
The code is syncronized, as we are dealing with multiple
threads, to provide logging of what have been done so it
can be read afterwards.

As files are converted they are moved right away to the
"done" directory, with this feature we aren't waiting until
the end of the execution.