[FileProcessing] Detect black borders (top-down)
With this change we will now be able to detect if
an image has black borders (currently top-down only).
The code is syncronized, as we are dealing with multiple
threads, to provide logging of what have been done so it
can be read afterwards.

As files are detected they are grouped to a directory,
labled the amount of pixel of the border in black, and
because of this we are waiting until the end of the
execution before relocating.

Note: For larger collections of images it is recommended
to use more than +50 threads as each image is analyzed one
pixel at a time on first the X axis and the Y axis. Complete
black (e.g. #000000) are matched, if not then the iteration
will end.