[Renfi] Coordinates to Axis
This change removes coordinates and instead
use standard Strings throughout the application
until the integer values are really needed as is.

Increased validation has also been implemented to
verify that the stated path is actually existing
before continuing with the flow.

Validation of Processing support has been refactored
to make it easier to read.