[FileProcessingUtilTest] Third batch of tests
For this change we have now verified that the
following methods are working:

1) renamingProcess - three cases that verifies a)
that multiple files can be renamed, b) that the
renaming process will continue even if one of the
files are missing or is incorrect, and c) that if
no history is available then it will just give a
message and continue the flow.

2) renamingUndoProcess - two cases that verifies a)
that multiple files can be restored, and b) the
selected files that were renamed are restored as well.

As these test cases were added I found a bug in the
createTargetDirectory method which has now been fixed,
the issue was that the return statements and the
condition itself was reversed for the intended purpose.

Finally, in this commit I also just made some minor
updates to remove some of the IDE recommendations.