[FileProcessingUtil] Compare resources refactoring
In this commit we have tried to optimize the compare
resources feature, not only in the overall runtime
but also in lowering the Cyclomatic complexity, and
in turn made the code easier to read by stating
better local variable names and less conditions.

Another change that was made is that you now need
to state the maximum percentage of difference that
the application should reach before discontinuing
the process.

The difference is based on per pixel for the entire
X row, once one row has been completed the difference
will be compared with the boundary limit and if the
limit is reached it will end. Obviously, by stating
less of a value the process of analyzing the files
will take less time but less of the file will be known.