[Exception] Exception handling introduced
To improve readability, secure functionality
and using exception handling correctly we are
now introducing these changes:

1) Introduced exception handling and added
some exception messages or corrected their
descriptions.

2) Corrected a few locations where the exception
message said one thing but was incorrectly verified
it in the if statement.

3) Moved source code into an if statement to
actually use the code during a positive flow, when
the negative flow will terminate the application.