[HelpGuideUtil] formatMessage refactoring
In this change we have now added the
formatMessage function that will append
strings to a template string, which was
priorly done by an overloaded function.

This new function is being used among
many printMessage and exception initializations,
and in turn replaced one of the overloaded
printMessage functions, as they serve the
same purpose.

Minor changes were also done to clean up
the code.