[Renfi] slf4j refactoring
In this change we have now added the slf4j
dependency to improve the logging of the
application. This will replace the printMessage
implementation that was previously made.

We have also introduced the LoggingUtil class
that will be storing the runtime process of
the application as the different threads are
processing the workload, regardless of the
feature. This was done to make it easier to
read the progress as each logging made by the
slf4j's dependency will all be scrambled / out
of order. Once the batch is done the entire map
of LoggingUtil will be printed to file.