[Renfi] Minor changes
In this change we have made some minor
changes as they were excluded from
previous commits:

1) Updating the Java version, in the build
section, as well as the version of said plugin.

2) Corrected the indentation which was excluded
from the previous commit.

3) Changed the order of the imports as the
Intellij settings were updated.

4) A bug was also found in the if-statement
where the file extensions are validated against
what is accepted.

5) As the slf4j logging was implemented this
change includes the removal of the printMessage
and the formatMessage functions from the
HelpGuideUtil class.