[FileProcessingUtil] moveResources refactoring
In this change we have now removed some code,
duplication, and moved the logic into a new
function called moveResources. It is used by
the Compare and by the Detect black borders
feature to move duplications, matching, and
modified resources. Other features do not use
this method as they are modified during runtime.