[Renfi] Minor changes
In this change we have made some minor
improvements to clean up the code and
in turn make it easier to read during
execution:

1) Removal of the ALLOWED_FLAGS arraylist
as we now verify that the respective flag
comes in the correct order and that it
uses one of the flags registered.

2) As we now want to validate the order
of the allowed flags, we have also moved
the path to be the 4th element in line, and
updated the HelpGuide in turn. The Boundary
element was also missing from the example
under the Operational Overview.

3) Several of the File processing flows
have been updated to have more descriptive
variables names and some even updated the
execution itself.

4) Minor changes made to the string constants.