[Renfi] Minor changes
In this change we have now made some minor
changes to improve readability:

1) Renaming and reordering of test constants,
one was also reworded.

2) Reverted the verification of the undoList
variable, to what it once was as the undoList
and the history shouldn't be equal, to make the
test case that verified this work once more.

3) Minor comments added for telling what happens
during the execution.

4) Logger introduced into the test classes,
instead of using System.